+++ bindings/wasm/src/lib.rs
@@ -137,11 +137,11 @@ export function inline(html: string, options?: InlineOptions): string;
 #[cfg(test)]
 pub mod tests {
     use super::*;
-    use wasm_bindgen_test::*;
+    use wasm_bindgen_test::wasm_bindgen_test;
 
     #[wasm_bindgen_test]
     fn default_config() {
-        let result = inline("<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1>Test</h1></body></html>", &JsValue::undefined()).unwrap();
+        let result = inline("<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1>Test</h1></body></html>", &JsValue::undefined()).expect("Inlines correctly");
         assert_eq!(result, "<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1 style=\"color:red;\">Test</h1></body></html>");
     }
 
@@ -149,10 +149,10 @@ pub mod tests {
     fn remove_style_tags() {
         let options = Options {
             remove_style_tags: true,
-            ..Default::default()
+            ..Options::default()
         };
-        let options = JsValue::from_serde(&options).unwrap();
-        let result = inline("<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1>Test</h1></body></html>", &options).unwrap();
+        let options = JsValue::from_serde(&options).expect("Valid value");
+        let result = inline("<html><head><title>Test</title><style>h1 { color:red; }</style></head><body><h1>Test</h1></body></html>", &options).expect("Inlines correctly");
         assert_eq!(result, "<html><head><title>Test</title></head><body><h1 style=\"color:red;\">Test</h1></body></html>");
     }
 }